Switches backing library to [kramdown][kd].
While Redcarpet[rc] would presumably have better rendering
performance, the library does not expose the internal tree
to renders. This fact made the sectioning/structure quite
difficult and awkward to construct... before considering
(nested) list nodes.

Upon starting work on nested lists, it became apparent that
Redcarpet (well, Sundown) rendering starts with leaf nodes.
That is rendering `ul > li > ol > li` begins at the `li`
within the `ol`. This fact makes it basically impossible
to determine the nesting structure, which is a show-stopper
for this library.

Redcarpet on the other hand presents a mechanism to walk
the children of nodes, starting at the root.

[kd]: https://github.com/gettalong/kramdown
[rc]: https://github.com/vmg/redcarpet